Fix the bug where on macOS and web that all moves have zero effect.
the reason things weren't working is because ReadLines
was not handling Unix-style file endings.

this commit also cleans up the code that computes
damage modifiers (i.e. are moves super effective, not very
effective, etc.)

+++ src/PokemonDemo.h
@@ -521,7 +521,7 @@ typedef struct
 	//I think I like these so I want to keep them
 	pokemon_database_data PokemonDatabase[POKEMON_DATABASE_LENGTH];
 	float NatureDatabase[25][6];
-	float TypeMatchupDatabase[18][18];
+	unsigned int TypeMatchupDatabase[18][18];
 	game_ui_scene UserInterfaces[256]; //typically we dafault using 256 elements
 	////////
